+++ src/class-setup.php
@@ -1,0 +1,83 @@
+<?php
+
+/**
+ * Class that defines plugin activation/deactivation/uninstall callbacks
+ *
+ * @link       http://example.com
+ * @since      1.0.0
+ *
+ * @package    PluginName
+ * @subpackage PluginName/src
+ *
+ * Heavily based on this Stack Exchange Q&A
+ * @author Franz Josef Kaiser/wecodemore
+ * @link   https://wordpress.stackexchange.com/questions/25910/uninstall-activate-deactivate-a-plugin-typical-features-how-to/25979#25979
+ */
+
+namespace Vendor_Name\Plugin_Name;
+
+class Setup {
+
+    protected static $instance;
+
+    public static function init() {
+        is_null( self::$instance ) AND self::$instance = new self;
+        return self::$instance;
+    }
+
+    public static function activate() {
+
+        // ddd( "Activate() is running" );
+
+        if ( ! current_user_can( 'activate_plugins' ) )
+            return;
+
+        // $requirements = new Requirements();
+        // $requirements->check();
+
+
+        $plugin = isset( $_REQUEST['plugin'] ) ? $_REQUEST['plugin'] : '';
+        check_admin_referer( "activate-plugin_{$plugin}" );
+
+
+        // register_custom_post_types();
+        // flush_rewrite_rules();
+
+        // Uncomment the following line to see the function in action
+        // exit( var_dump( $_GET ) );
+    }
+
+    public static function deactivate() {
+        if ( ! current_user_can( 'activate_plugins' ) )
+            return;
+
+        // This feels like a workaround. Produces "Are you sure you want to do this?" error if deactivated due to failed requirements check.    
+        if( Requirements::requirements_met() ) {
+            $plugin = isset( $_REQUEST['plugin'] ) ? $_REQUEST['plugin'] : '';
+            check_admin_referer( "deactivate-plugin_{$plugin}" );
+        }
+
+        // @link https://knowthecode.io/labs/custom-post-type-basics/episode-8
+        // flush_rewrite_rules() results in weird behavior. Use this instead...
+        // delete_option( 'rewrite_rules' );
+
+        // Uncomment the following line to see the function in action
+        // exit( var_dump( $_GET ) );
+    }
+
+    public static function uninstall() {
+        if ( ! current_user_can( 'activate_plugins' ) )
+            return;
+        check_admin_referer( 'bulk-plugins' );
+
+        // Important: Check if the file is the one
+        // that was registered during the uninstall hook.
+        if ( __FILE__ != WP_UNINSTALL_PLUGIN )
+            return;
+
+        // Remove files and database tables here
+
+        # Uncomment the following line to see the function in action
+        # exit( var_dump( $_GET ) );
+    }
+}