+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## June 21, 2017
+
+* Added Setup class for activation/deactivation/uninstall processes
+* Added system requirements check
+
 ## June 20, 2017
 
 * Fork of the WordPress Plugin Boilerplate NS.